CommunicationUtilities: Move Trace method implementations to Framework
- Move the implementations of the CommunicationUtilities.Trace methods to FrameworkCommunicationUtilities in Microsoft.Build.Framework.
- Leave the CommunicationUtilities.Trace methods in place but delegate to the methods in FrameworkCommunicationUtilities.
