+++ src/Shared/CommunicationsUtilities.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
-using System.IO;
 using System.IO.Pipes;
 
 #if RUNTIME_TYPE_NETCORE
@@ -19,12 +18,13 @@ using System.Security.Principal;
 using System.Reflection;
 using System.Security.Cryptography;
 using System.Text;
-using System.Threading;
+using Microsoft.Build.BackEnd;
 using Microsoft.Build.Framework;
 using Microsoft.Build.Shared;
-using Microsoft.Build.BackEnd;
 
 #if !FEATURE_APM
+using System.IO;
+using System.Threading;
 using System.Threading.Tasks;
 #endif
 
@@ -413,26 +413,6 @@ namespace Microsoft.Build.Internal
         private const int DefaultNodeConnectionTimeout = 900 * 1000; // 15 minutes; enough time that a dev will typically do another build in this time
 
         /// <summary>
-        /// Whether to trace communications
-        /// </summary>
-        private static readonly bool s_trace = Traits.Instance.DebugNodeCommunication;
-
-        /// <summary>
-        /// Lock trace to ensure we are logging in serial fashion.
-        /// </summary>
-        private static readonly LockType s_traceLock = new LockType();
-
-        /// <summary>
-        /// Place to dump trace
-        /// </summary>
-        private static string s_debugDumpPath;
-
-        /// <summary>
-        /// Ticks at last time logged
-        /// </summary>
-        private static long s_lastLoggedTicks = DateTime.UtcNow.Ticks;
-
-        /// <summary>
         /// Delegate to debug the communication utilities.
         /// </summary>
         internal delegate void LogDebugCommunications(string format, params object[] stuff);
@@ -798,130 +778,52 @@ namespace Microsoft.Build.Internal
         /// Writes trace information to a log file
         /// </summary>
         internal static void Trace<T>(string format, T arg0)
-        {
-            Trace(nodeId: -1, format, arg0);
-        }
+            => FrameworkCommunicationsUtilities.Trace(format, arg0);
 
         /// <summary>
         /// Writes trace information to a log file
         /// </summary>
         internal static void Trace<T>(int nodeId, string format, T arg0)
-        {
-            if (s_trace)
-            {
-                TraceCore(nodeId, string.Format(format, arg0));
-            }
-        }
+            => FrameworkCommunicationsUtilities.Trace(nodeId, format, arg0);
 
         /// <summary>
         /// Writes trace information to a log file
         /// </summary>
         internal static void Trace<T0, T1>(string format, T0 arg0, T1 arg1)
-        {
-            Trace(nodeId: -1, format, arg0, arg1);
-        }
+            => FrameworkCommunicationsUtilities.Trace(format, arg0, arg1);
 
         /// <summary>
         /// Writes trace information to a log file
         /// </summary>
         internal static void Trace<T0, T1>(int nodeId, string format, T0 arg0, T1 arg1)
-        {
-            if (s_trace)
-            {
-                TraceCore(nodeId, string.Format(format, arg0, arg1));
-            }
-        }
+            => FrameworkCommunicationsUtilities.Trace(nodeId, format, arg0, arg1);
 
         /// <summary>
         /// Writes trace information to a log file
         /// </summary>
         internal static void Trace<T0, T1, T2>(string format, T0 arg0, T1 arg1, T2 arg2)
-        {
-            Trace(nodeId: -1, format, arg0, arg1, arg2);
-        }
+            => FrameworkCommunicationsUtilities.Trace(format, arg0, arg1, arg2);
 
         /// <summary>
         /// Writes trace information to a log file
         /// </summary>
         internal static void Trace<T0, T1, T2>(int nodeId, string format, T0 arg0, T1 arg1, T2 arg2)
-        {
-            if (s_trace)
-            {
-                TraceCore(nodeId, string.Format(format, arg0, arg1, arg2));
-            }
-        }
+            => FrameworkCommunicationsUtilities.Trace(nodeId, format, arg0, arg1, arg2);
 
         /// <summary>
         /// Writes trace information to a log file
         /// </summary>
         internal static void Trace(string format, params object[] args)
-        {
-            Trace(nodeId: -1, format, args);
-        }
+            => FrameworkCommunicationsUtilities.Trace(format, args);
 
         /// <summary>
         /// Writes trace information to a log file
         /// </summary>
         internal static void Trace(int nodeId, string format, params object[] args)
-        {
-            if (s_trace)
-            {
-                string message = string.Format(CultureInfo.CurrentCulture, format, args);
-                TraceCore(nodeId, message);
-            }
-        }
+            => FrameworkCommunicationsUtilities.Trace(nodeId, format, args);
 
         internal static void Trace(int nodeId, string message)
-        {
-            if (s_trace)
-            {
-                TraceCore(nodeId, message);
-            }
-        }
-
-        /// <summary>
-        /// Writes trace information to a log file
-        /// </summary>
-        private static void TraceCore(int nodeId, string message)
-        {
-            lock (s_traceLock)
-            {
-                s_debugDumpPath ??= FrameworkDebugUtils.DebugPath;
-
-                if (String.IsNullOrEmpty(s_debugDumpPath))
-                {
-                    s_debugDumpPath = FileUtilities.TempFileDirectory;
-                }
-                else
-                {
-                    Directory.CreateDirectory(s_debugDumpPath);
-                }
-
-                try
-                {
-                    string fileName = @"MSBuild_CommTrace_PID_{0}";
-                    if (nodeId != -1)
-                    {
-                        fileName += "_node_" + nodeId;
-                    }
-
-                    fileName += ".txt";
-
-                    using (StreamWriter file = FileUtilities.OpenWrite(
-                        string.Format(CultureInfo.CurrentCulture, Path.Combine(s_debugDumpPath, fileName), EnvironmentUtilities.CurrentProcessId, nodeId), append: true))
-                    {
-                        long now = DateTime.UtcNow.Ticks;
-                        float millisecondsSinceLastLog = (float)(now - s_lastLoggedTicks) / 10000L;
-                        s_lastLoggedTicks = now;
-                        file.WriteLine("{0} (TID {1}) {2,15} +{3,10}ms: {4}", Thread.CurrentThread.Name, Thread.CurrentThread.ManagedThreadId, now, millisecondsSinceLastLog, message);
-                    }
-                }
-                catch (IOException)
-                {
-                    // Ignore
-                }
-            }
-        }
+            => FrameworkCommunicationsUtilities.Trace(nodeId, message);
 
         /// <summary>
         /// Gets a hash code for this string.  If strings A and B are such that A.Equals(B), then
